refactor: CommandFactory 관련 클래스를 새 패키지로 묶음
