+++ src/main/java/com/nubble/backend/post/comment/service/CommentService.java
@@ -3,6 +3,7 @@ package com.nubble.backend.post.comment.service;
 import com.nubble.backend.post.comment.domain.Comment;
 import com.nubble.backend.post.comment.service.CommentCommand.CommentCreateCommand;
 import com.nubble.backend.post.comment.service.CommentCommand.CommentDeleteCommand;
+import com.nubble.backend.post.comment.service.factory.CommentFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
